Map fragment
God has enlightened me.  I see no one, I fear no one.
Jokes aside, I've implemented the fragment of the map that you access when you click the icon on the bottom navigation bar. It displays the markers, and it shows info if you click one of them. The only little problem would be that it is not actualized when you add a new post, unless you restart the app. If you are Mateo, contact me if you want me to try to solve this.